backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pyodbc==5.0.1
plaid-python==9.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2